python-api/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles==23.2.1
huggingface_hub>=0.28.0
requests==2.31.0
pydantic==2.5.0